Particle Systems Prototype
This commit contains the Particle Systems Prototype for the Shred
prototypes project. This is the first commit and is temporarily the
master. Eventually, the master will contain a skeleton for the main menu
UI (not functioning) with buttons to load each of the five prototypes as
a scene for demonstration.  Ideally the master would contain the main
menu UI and each developer could branch the master to add in their
prototype and have it pulled to the master.